# openapi.yaml
...
openapi: 3.0.3
info:
    title: Go-Kratos Console API
    description: Management console for Go-Kratos
    contact:
        name: Sendya
        url: https://github.com/sendya
        email: yladmxa@gmail.com
    license:
        name: Apache License
        url: https://github.com/google/gnostic/blob/master/LICENSE
    version: 1.0.0
paths:
...
    /api/console/zone:
        get:
            tags:
                - Zone
            description: |-
                List returns a list of zones.

                 The list is sorted by zone id.
            operationId: Zone_List
            parameters:
                - name: pagination.current
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.page_size
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.total
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.raw_total
                  in: query
                  schema:
                    type: string
                - name: env
                  in: query
                  description: 查询条件 env 环境变量
                  schema:
                    type: string
                - name: status
                  in: query
                  description: 查询条件 status 状态 1=正常，2=禁用
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.GetZoneListReply'
        post:
            tags:
                - Zone
            description: |-
                Create creates a new zone.

                 ...
            operationId: Zone_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.console.resource.CreateZoneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.CreateZoneReply'
...
    /api/console/zone/{id}:
        get:
            tags:
                - Zone
            description: |-
                Get returns the zone.

                 Query parameter "id" is required.
                 Example: /api/console/zone/1
            operationId: Zone_Get
            parameters:
                - name: id
                  in: path
                  description: 可用区ID
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.GetZoneReply'
        put:
            tags:
                - Zone
            description: |-
                Update updates the zone.

                 ...
            operationId: Zone_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.console.resource.UpdateZoneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.UpdateZoneReply'
        delete:
            tags:
                - Zone
            description: |-
                Delete deletes the zone.

                 ...
            operationId: Zone_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.DeleteZoneReply'
    /api/console/zone/{id}/disable:
        post:
            tags:
                - Zone
            description: |-
                Disable disables the zone.

                 ...
            operationId: Zone_Disable
            parameters:
                - name: id
                  in: path
                  description: 可用区ID
                  required: true
                  schema:
                    type: integer
                    format: int32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.console.resource.DisableZoneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.DisableZoneReply'
components:
    schemas:
...
        api.console.resource.CreateZoneReply:
            type: object
            properties:
                data:
                    allOf:
                        - $ref: '#/components/schemas/api.console.resource.ZoneInfo'
                    description: 可用区
        api.console.resource.CreateZoneRequest:
            type: object
            properties:
                name:
                    type: string
                    description: 可用区名称
                code:
                    type: string
                    description: 可用区代码
                region_name:
                    type: string
                    description: 地区名称
                region_code:
                    type: string
                    description: 地区代码
                env:
                    type: string
                    description: 环境
        api.console.resource.DeleteZoneReply:
            type: object
            properties: {}
        api.console.resource.DisableZoneReply:
            type: object
            properties: {}
        api.console.resource.DisableZoneRequest:
            type: object
            properties:
                id:
                    type: integer
                    description: 可用区ID
                    format: int32
                remark:
                    type: string
                    description: 备注
        api.console.resource.GetZoneListReply:
            type: object
            properties:
                pagination:
                    $ref: '#/components/schemas/protobuf.Pagination'
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.console.resource.ZoneInfo'
                    description: 可用区列表
        api.console.resource.GetZoneReply:
            type: object
            properties:
                data:
                    allOf:
                        - $ref: '#/components/schemas/api.console.resource.ZoneInfo'
                    description: 可用区
...
        api.console.resource.UpdateZoneReply:
            type: object
            properties: {}
        api.console.resource.UpdateZoneRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                name:
                    type: string
                    description: 可用区名称
                code:
                    type: string
                    description: 可用区代码
                region_name:
                    type: string
                    description: 地区名称
                region_code:
                    type: string
                    description: 地区代码
                env:
                    type: string
                    description: 环境
        api.console.resource.ZoneInfo:
            type: object
            properties:
                id:
                    type: integer
                    description: 可用区ID
                    format: int32
                name:
                    type: string
                    description: 可用区名称
                code:
                    type: string
                    description: 可用区代码
                region_name:
                    type: string
                    description: 地区名称
                region_code:
                    type: string
                    description: 地区代码
                env:
                    type: string
                    description: 环境
                status:
                    type: integer
                    description: 状态
                    format: enum
                created_at:
                    type: string
                    description: 创建时间
                updated_at:
                    type: string
                    description: 更新时间
            description: 可用区
        protobuf.Pagination:
            type: object
            properties:
                current:
                    type: integer
                    format: int32
                page_size:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                raw_total:
                    type: string
            description: 通用的分页信息
    securitySchemes:
        ApiKeyAuth:
            type: apiKey
            name: Authorization
            in: header
...